requirements.txt
discord.py
aiomysql
aiosmtplib